# api-docs.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
paths:
...
  /api/trainings/{id}:
    put:
      tags:
      - training-rest-controller
      summary: Replace a training
      operationId: updateTraining
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrainingDTO"
        required: true
      responses:
        "200":
          description: Training replaced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingDTO"
        "400":
          description: Bad request - Invalid parameters
        "401":
          description: Unauthorized access - Authentication is required
        "403":
          description: Forbidden - You don't have permission to access
        "404":
          description: Training not found
        "500":
          description: Internal server error
    delete:
      tags:
      - training-rest-controller
      summary: Delete a training
      operationId: deleteTraining
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Training deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingDTO"
        "400":
          description: Bad request - Invalid parameters
        "401":
          description: Unauthorized access - Authentication is required
        "403":
          description: Forbidden - You don't have permission to access
        "404":
          description: Training not found
        "500":
          description: Internal server error
...
  /api/trainings/unsubscribed/{id}:
    post:
      tags:
      - training-rest-controller
      summary: Unsubscribe a user from a training
      operationId: unsubscribeTraining
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: User unsubscribed from training successfully
        "400":
          description: Bad request - Invalid input data
        "401":
          description: Unauthorized access
        "404":
          description: Training or User not found
        "500":
          description: Internal server error
  /api/trainings/subscribed/{id}:
    post:
      tags:
      - training-rest-controller
      operationId: subscribeTraining
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /api/trainings/:
    post:
      tags:
      - training-rest-controller
      summary: Create a training
      operationId: createTraining
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrainingDTO"
        required: true
      responses:
        "201":
          description: Training created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingDTO"
        "400":
          description: Bad request - Invalid parameters
        "401":
          description: Unauthorized access - Authentication is required
        "403":
          description: Forbidden - You don't have permission to access
        "500":
          description: Internal server error
...
  /api/foods/:
    post:
      tags:
      - food-rest-controller
      summary: Create a food
      operationId: createTraining_1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FoodDTO"
        required: true
      responses:
        "201":
          description: Food created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FoodDTO"
        "400":
          description: Bad request - Invalid parameters
        "401":
          description: Unauthorized access - Authentication is required
        "403":
          description: Forbidden - You don't have permission to access
        "500":
          description: Internal server error
...
  /api/trainings/{trainingId}:
    get:
      tags:
      - training-rest-controller
      summary: Get a training by ID
      operationId: getTrainingByIdd_1
      parameters:
      - name: trainingId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Training found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingDTO"
        "400":
          description: Bad request - Invalid parameters
        "401":
          description: Unauthorized access - Authentication is required
        "403":
          description: Forbidden - You don't have permission to access
        "404":
          description: Training not found
        "500":
          description: Internal server error
  /api/trainings/myTrainings:
    get:
      tags:
      - training-rest-controller
      summary: Get all trainings of user
      operationId: getUserTrainings
      responses:
        "200":
          description: Found all trainings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrainingDTO"
        "400":
          description: Bad request - Invalid parameters
        "401":
          description: Unauthorized access - Authentication is required
        "403":
          description: Forbidden - You don't have permission to access
        "404":
          description: Trainings not found
        "500":
          description: Internal server error
...
components:
  schemas:
...
    TrainingDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        intensity:
          type: string
        duration:
          type: integer
          format: int32
        exercises:
          type: string
        goal:
          type: string
        userId:
          type: integer
          format: int64
...
    FoodDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        calories:
          type: integer
          format: int32
        type:
          type: string
        listIdsNutrition:
          type: array
          items:
            type: integer
            format: int64
...